[SYS-3275] Igore directories and errors during local directory cleaning up (#205)
Local directory cleaning up during sanitization fails if there is directory locally. We changed the behavior here so that:

Local directories are ignored
File deletion errors are ignored as well. This is ok since we fetch the latest CM/M from s3 when resync_on_open=True

+++ cloud/cloud_env_impl.cc
@@ -1706,9 +1706,32 @@ Status CloudEnvImpl::SanitizeDirectory(const DBOptions& options,
       continue;
     }
     std::string pathname = local_name + "/" + file.name;
-    st = env->DeleteFile(pathname);
+    bool is_dir;
+    st = env->IsDirectory(pathname, &is_dir);
     if (!st.ok()) {
-      return st;
+      Log(InfoLogLevel::WARN_LEVEL, info_log_,
+          "[cloud_env_impl] IsDirectory fails for %s %s", pathname.c_str(),
+          st.ToString().c_str());
+      // file checking failure ignored
+      continue;
+    }
+
+    // Directories are ignored
+    if (is_dir) {
+      Log(InfoLogLevel::INFO_LEVEL, info_log_,
+          "[cloud_env_impl] Not deleting: %s since it's a directory",
+          pathname.c_str());
+      continue;
+    }
+
+    st = env->DeleteFile(pathname);
+    TEST_SYNC_POINT_CALLBACK("CloudEnvImpl::SanitizeDirectory:AfterDeleteFile", &st);
+    if (!st.ok()) {
+      Log(InfoLogLevel::WARN_LEVEL, info_log_,
+          "[cloud_env_impl] DeleteFile: %s fails %s", pathname.c_str(),
+          st.ToString().c_str());
+      // local file cleaning up failure will be ignored
+      continue;
     }
     Log(InfoLogLevel::INFO_LEVEL, info_log_,
         "[cloud_env_impl] SanitizeDirectory cleaned-up: '%s'",
@@ -1716,10 +1739,8 @@ Status CloudEnvImpl::SanitizeDirectory(const DBOptions& options,
   }
 
   if (!st.ok()) {
-    Log(InfoLogLevel::DEBUG_LEVEL, info_log_,
-        "[cloud_env_impl] SanitizeDirectory error opening dir %s %s",
-        local_name.c_str(), st.ToString().c_str());
-    return st;
+    // ignore all the errors generated during cleaning up
+    st = Status::OK();
   }
 
   Log(InfoLogLevel::DEBUG_LEVEL, info_log_,
